% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx2dfs.R
\name{withNames}
\alias{withNames}
\title{Helper function for more convenient input (sheet name, data frame, sheet name, data frame, ...).}
\usage{
withNames(...)
}
\arguments{
\item{...}{alterning arguments: sheet name 1, data frame 1, sheet name 2, data frame 2, ...}
}
\value{
A list of the data frames with the names given each before.
}
\description{
Helper function for more convenient input (sheet name, data frame, sheet name, data frame, ...).
}
\examples{
df1 <- data.frame(A=c(1, 2), B=c(3, 4))
df2 <- data.frame(C=c(5, 6), D=c(7, 8))
xlsx_fpath <- file.path(tempdir(), "testout.xlsx")
dfs2xlsx(withNames("sheet1", df1, "sheet2", df2), xlsx_fpath)
file.remove(xlsx_fpath)
}
